[EI-260] [web] refactor: Update useSplitIndicatorBoard hook to manage split screen functionality in indicator board page
